put table in seperate component
